+++ CHANGELOG.md
@@ -8,10 +8,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Datagrid cells can now be shaded with a color gradient (https://github.com/PhenoApps/Field-Book/pull/1488)
+- Datagrid columns can now be sorted (https://github.com/PhenoApps/Field-Book/pull/1488)
+- Multiple attributes can now be selected as row headers in Datagrid (https://github.com/PhenoApps/Field-Book/pull/1488)
+- Datagrid column width can be adjusted to match values (https://github.com/PhenoApps/Field-Book/pull/1488)
+- Datagrid columns can now be locked to remain visible while scrolling (https://github.com/PhenoApps/Field-Book/pull/1488)
+- New Datagrid Map View matches spatial layout of the field (https://github.com/PhenoApps/Field-Book/pull/1488)
+- Multiple BrAPI accounts can now be added (https://github.com/PhenoApps/Field-Book/pull/1412)
+- BrAPI accounts are now managed in Android system settings (https://github.com/PhenoApps/Field-Book/pull/1412)
 
 ### Changed
 
 ### Fixed
+
+## [v7.2.3] - 2026-07-24
+
+### Fixed
+- Categorical traits with multiple categories selected now correctly reset (https://github.com/PhenoApps/Field-Book/pull/1498)
+- Auto advance in date traits now correctly resets after a value is saved (https://github.com/PhenoApps/Field-Book/pull/1496)
+
+## [v7.2.2] - 2026-07-07
+
+### Fixed
+- Photos are no longer corrupted on certain Samsung devices (https://github.com/PhenoApps/Field-Book/pull/1489)
+- Numerous BrAPI bug fixes and improvements (https://github.com/PhenoApps/Field-Book/pull/1483)
+
+## [v7.2.1] - 2026-04-30
+
+### Added
+- New Experimental Traits setting to show traits under development (https://github.com/PhenoApps/Field-Book/pull/1469)
+
+### Fixed
+- Large values now fit correctly in the repeated values view (https://github.com/PhenoApps/Field-Book/pull/1468)
 - InfoBars no longer disappear when attempting to display trait data (https://github.com/PhenoApps/Field-Book/pull/1458)
 - Repeated measures can now still be collected when data input is frozen (https://github.com/PhenoApps/Field-Book/pull/1459)
 - Nix trait now works correctly in public builds (https://github.com/PhenoApps/Field-Book/pull/1465)
@@ -934,3 +962,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [v7.1.2]: https://github.com/PhenoApps/Field-Book/releases/tag/7.1.2
 
 [v7.2.0]: https://github.com/PhenoApps/Field-Book/releases/tag/7.2.0
+
+[v7.2.1]: https://github.com/PhenoApps/Field-Book/releases/tag/7.2.1
+
+[v7.2.2]: https://github.com/PhenoApps/Field-Book/releases/tag/7.2.2
+[v7.2.3]: https://github.com/PhenoApps/Field-Book/releases/tag/7.2.3